requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
camelot-py==0.11.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==7.1.0
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.14
dirtyjson==1.0.8
docx==0.2.4
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.2
filelock==3.16.0
fitz==0.0.1.dev2
fonttools==4.53.1
frontend==0.0.3
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.0
grpcio==1.66.1
grpcio-tools==1.66.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
hyperframe==6.0.1
idna==3.9
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
langchain==0.2.16
langchain-community==0.2.17
langchain-core==0.2.40
langchain-text-splitters==0.2.4
langsmith==0.1.120
llama-index-core==0.11.9
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.8.4
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.8.0
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
nltk==3.9.1
numpy==1.26.4
openpyxl==3.1.5
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
pdfminer.six==20240706
pillow==10.4.0
portalocker==2.10.1
#PropBots==0.0.1
protobuf==5.28.1
prov==2.0.1
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pydeck==0.9.1
pydot==3.0.1
Pygments==2.18.0
pyparsing==3.1.4
pypdf==4.3.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
qdrant-client==1.11.1
rdflib==6.3.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
scipy==1.14.1
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
streamlit==1.36.0
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.7.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traits==6.3.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
wrapt==1.16.0
yarl==1.11.1
pymupdf
PyMuPDF
opencv-python
llama-index
llama-index-vector-stores-qdrant
llama-index-vector-stores
fastembed
git+https://github.com/openai/CLIP.git

- e.
